Implementation of core function for lib poisson_1D : In this commit we implemented :
set_GB_opertor_colMajor_poisson1D_ID :
this function return an identity matrix in the format GB
for our poisson problem.

We implemented set_dense_RHS_DBS_1D:

This function initialize the right hand part of equation in our case b using
dirichilet boundary condition BC0 and BC1  :
We set to zero except on the boundaries of the domain.

We implemented set_analytical_solution_DBC_1D:

as the function that return the analytical solution to the problem.
We will later use it to compute the relative error of our model.

We implemented the function set_grid_points_1D :

This function intialize the x vector as a grid of uniformly spaced point.
We assume L=1 as this is given in the previous question of the TD q1
"the domain span from 0 to 1 "