Implementation of relative_forward_error & indexABCol
In this commit we implemented 2 function :

relative_forward_error :
this function calculate the forward error between our prediction and the exact solution.

indexAB :
In this function we calculate the colmajor index of corresponding to the entry i,j in
a tridiag matrix stored in GB format  .
because we are in the tridiagonal case we only have to treat 3 case :
upper diag  j == i + 1
main diag  j==i
lower diag j== i-1

this is define using the proprety of tridiag matrices.